fix: optionalize health check to support global NEG backend services
Create a new local `backend_with_healthchecks` that represents all non global NEG backend services which are iterated over (as opposed to the full map of backends) to create the health check resources.

Furthermore if the value of `health_check` is null, an empty array is passed to the `compute_backend_service`'s `health_checks` argument.

Signed-off-by: Naseem <naseem@transit.app>