Serverless NEGs submodule
This introduces a `serverless_negs` submodule.

What's specific about this is that it "disables" a bunch of the functionality
from the main template which was written with GCE in mind.

Many such concepts (e.g. firewall rules, instance groups, health checks) don't
apply to serverless NEGs. Similarly such "backend" fields that don't apply to
SNEGs are disabled as well.

The resulting "serverless_negs" module is easier to use and encourages use of
managed certificates and http-to-https redirects (which are recommended for
serverless user persona who are more inclined to run websites).

I have not run the tests myself, so they are purely copy-paste.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>