Fix the limits and requests blocks.
It was resulting in the following errors from "terraform validate":

Error: Unsupported argument

  on main.tf line 121, in resource "kubernetes_replication_controller" "nginx":
 121:             limits = {

An argument named "limits" is not expected here. Did you mean to define a
block of type "limits"?

Error: Unsupported argument

  on main.tf line 126, in resource "kubernetes_replication_controller" "nginx":
 126:             requests = {

An argument named "requests" is not expected here. Did you mean to define a
block of type "requests"?

Signed-off-by: Kunal Gangakhedkar <kunalg@hotstar.com>