fix: Create the `random_id.certificate` resource only when needed. (#184)
The option to add a random prefix to the name of the managed certificate
was added as part of #160. When this option is disabled, the `random_id`
resource is still created, despite not being used anywhere. This could
cause confusion when looking at the output of `terraform plan`.

This PR makes the creation of that resource conditional on the
`random_certificate_suffix` flag. When the flag is set to `false`, v5.0
and v6.0 generate the same resources.