Expand url map functionality in serverless_negs
Previously, this module provided the option to either construct a
trivial url map which maps everything to a single backend OR pass in
a url map. The problem with passing in a url map is that the url map
depends on the backend services that are created in the serverless_negs
module which makes this approach unworkable.

This work adds the input variable `url_map_spec` which allows the
user to pass in a map of objects which are keyed on the same set
as the keys of the `backends` input variable.

This new variable enables the addition of host_rule/path_matcher pairs
to the google_compute_url_map block with a list of path_rules for each
of these pairs.