fix(url_map): Adding create_before_destroy lifecycle.
The `target_http_proxy` resource depends on the `google_compute_url_map`
resource, so Terraform attempts to update the `url_map` first when
changes occur. If a custom `url_map` is added, the module tries to
delete the default `url_map` before updating the one the `https_proxy`
relies on, resulting in an error.